idpf: Refactor filter operations in separate file
Added new files 'idpf_fltr.c' and 'idpf_fltr.h'.

This patch makes an effort to move all filter operations into a
separate file. This reduces the burden on other files where these
operations were clubbed together, and makes them look clean now.

Also, moved some of the structures from idpf_virtchnl.c to
idpf_virtchnl.h

There is no change in the code or functionality before and after
the change.
