+++ drivers/net/ethernet/intel/idpf/idpf_lib.c
@@ -2,6 +2,7 @@
 /* Copyright (C) 2023 Intel Corporation */
 
 #include "idpf.h"
+#include "idpf_fltr.h"
 #include "idpf_virtchnl.h"
 
 static const struct net_device_ops idpf_netdev_ops_splitq;
@@ -419,293 +420,6 @@ send_dealloc_vecs:
 	idpf_send_dealloc_vectors_msg(adapter);
 
 	return err;
-}
-
-/**
- * idpf_find_mac_filter - Search filter list for specific mac filter
- * @vconfig: Vport config structure
- * @macaddr: The MAC address
- *
- * Returns ptr to the filter object or NULL. Must be called while holding the
- * mac_filter_list_lock.
- **/
-static struct idpf_mac_filter *idpf_find_mac_filter(struct idpf_vport_config *vconfig,
-						    const u8 *macaddr)
-{
-	struct idpf_mac_filter *f;
-
-	if (!macaddr)
-		return NULL;
-
-	list_for_each_entry(f, &vconfig->user_config.mac_filter_list, list) {
-		if (ether_addr_equal(macaddr, f->macaddr))
-			return f;
-	}
-
-	return NULL;
-}
-
-/**
- * __idpf_del_mac_filter - Delete a MAC filter from the filter list
- * @vport_config: Vport config structure
- * @macaddr: The MAC address
- *
- * Returns 0 on success, error value on failure
- **/
-static int __idpf_del_mac_filter(struct idpf_vport_config *vport_config,
-				 const u8 *macaddr)
-{
-	struct idpf_mac_filter *f;
-
-	spin_lock_bh(&vport_config->mac_filter_list_lock);
-	f = idpf_find_mac_filter(vport_config, macaddr);
-	if (f) {
-		list_del(&f->list);
-		kfree(f);
-	}
-	spin_unlock_bh(&vport_config->mac_filter_list_lock);
-
-	return 0;
-}
-
-/**
- * idpf_del_mac_filter - Delete a MAC filter from the filter list
- * @vport: Main vport structure
- * @np: Netdev private structure
- * @macaddr: The MAC address
- * @async: Don't wait for return message
- *
- * Removes filter from list and if interface is up, tells hardware about the
- * removed filter.
- **/
-static int idpf_del_mac_filter(struct idpf_vport *vport,
-			       struct idpf_netdev_priv *np,
-			       const u8 *macaddr, bool async)
-{
-	struct idpf_vport_config *vport_config;
-	struct idpf_mac_filter *f;
-
-	vport_config = np->adapter->vport_config[np->vport_idx];
-
-	spin_lock_bh(&vport_config->mac_filter_list_lock);
-	f = idpf_find_mac_filter(vport_config, macaddr);
-	if (f) {
-		f->remove = true;
-	} else {
-		spin_unlock_bh(&vport_config->mac_filter_list_lock);
-
-		return -EINVAL;
-	}
-	spin_unlock_bh(&vport_config->mac_filter_list_lock);
-
-	if (np->state == __IDPF_VPORT_UP) {
-		int err;
-
-		err = idpf_add_del_mac_filters(vport, np, false, async);
-		if (err)
-			return err;
-	}
-
-	return  __idpf_del_mac_filter(vport_config, macaddr);
-}
-
-/**
- * __idpf_add_mac_filter - Add mac filter helper function
- * @vport_config: Vport config structure
- * @macaddr: Address to add
- *
- * Takes mac_filter_list_lock spinlock to add new filter to list.
- */
-static int __idpf_add_mac_filter(struct idpf_vport_config *vport_config,
-				 const u8 *macaddr)
-{
-	struct idpf_mac_filter *f;
-
-	spin_lock_bh(&vport_config->mac_filter_list_lock);
-
-	f = idpf_find_mac_filter(vport_config, macaddr);
-	if (f) {
-		f->remove = false;
-		spin_unlock_bh(&vport_config->mac_filter_list_lock);
-
-		return 0;
-	}
-
-	f = kzalloc(sizeof(*f), GFP_ATOMIC);
-	if (!f) {
-		spin_unlock_bh(&vport_config->mac_filter_list_lock);
-
-		return -ENOMEM;
-	}
-
-	ether_addr_copy(f->macaddr, macaddr);
-	list_add_tail(&f->list, &vport_config->user_config.mac_filter_list);
-	f->add = true;
-
-	spin_unlock_bh(&vport_config->mac_filter_list_lock);
-
-	return 0;
-}
-
-/**
- * idpf_add_mac_filter - Add a mac filter to the filter list
- * @vport: Main vport structure
- * @np: Netdev private structure
- * @macaddr: The MAC address
- * @async: Don't wait for return message
- *
- * Returns 0 on success or error on failure. If interface is up, we'll also
- * send the virtchnl message to tell hardware about the filter.
- **/
-static int idpf_add_mac_filter(struct idpf_vport *vport,
-			       struct idpf_netdev_priv *np,
-			       const u8 *macaddr, bool async)
-{
-	struct idpf_vport_config *vport_config;
-	int err;
-
-	vport_config = np->adapter->vport_config[np->vport_idx];
-	err = __idpf_add_mac_filter(vport_config, macaddr);
-	if (err)
-		return err;
-
-	if (np->state == __IDPF_VPORT_UP)
-		err = idpf_add_del_mac_filters(vport, np, true, async);
-
-	return err;
-}
-
-/**
- * idpf_del_all_mac_filters - Delete all MAC filters in list
- * @vport: main vport struct
- *
- * Takes mac_filter_list_lock spinlock.  Deletes all filters
- */
-static void idpf_del_all_mac_filters(struct idpf_vport *vport)
-{
-	struct idpf_vport_config *vport_config;
-	struct idpf_mac_filter *f, *ftmp;
-
-	vport_config = vport->adapter->vport_config[vport->idx];
-	spin_lock_bh(&vport_config->mac_filter_list_lock);
-
-	list_for_each_entry_safe(f, ftmp, &vport_config->user_config.mac_filter_list,
-				 list) {
-		list_del(&f->list);
-		kfree(f);
-	}
-
-	spin_unlock_bh(&vport_config->mac_filter_list_lock);
-}
-
-/**
- * idpf_restore_mac_filters - Re-add all MAC filters in list
- * @vport: main vport struct
- *
- * Takes mac_filter_list_lock spinlock.  Sets add field to true for filters to
- * resync filters back to HW.
- */
-static void idpf_restore_mac_filters(struct idpf_vport *vport)
-{
-	struct idpf_vport_config *vport_config;
-	struct idpf_mac_filter *f;
-
-	vport_config = vport->adapter->vport_config[vport->idx];
-	spin_lock_bh(&vport_config->mac_filter_list_lock);
-
-	list_for_each_entry(f, &vport_config->user_config.mac_filter_list, list)
-		f->add = true;
-
-	spin_unlock_bh(&vport_config->mac_filter_list_lock);
-
-	idpf_add_del_mac_filters(vport, netdev_priv(vport->netdev),
-				 true, false);
-}
-
-/**
- * idpf_remove_mac_filters - Remove all MAC filters in list
- * @vport: main vport struct
- *
- * Takes mac_filter_list_lock spinlock. Sets remove field to true for filters
- * to remove filters in HW.
- */
-static void idpf_remove_mac_filters(struct idpf_vport *vport)
-{
-	struct idpf_vport_config *vport_config;
-	struct idpf_mac_filter *f;
-
-	vport_config = vport->adapter->vport_config[vport->idx];
-	spin_lock_bh(&vport_config->mac_filter_list_lock);
-
-	list_for_each_entry(f, &vport_config->user_config.mac_filter_list, list)
-		f->remove = true;
-
-	spin_unlock_bh(&vport_config->mac_filter_list_lock);
-
-	idpf_add_del_mac_filters(vport, netdev_priv(vport->netdev),
-				 false, false);
-}
-
-/**
- * idpf_deinit_mac_addr - deinitialize mac address for vport
- * @vport: main vport structure
- */
-static void idpf_deinit_mac_addr(struct idpf_vport *vport)
-{
-	struct idpf_vport_config *vport_config;
-	struct idpf_mac_filter *f;
-
-	vport_config = vport->adapter->vport_config[vport->idx];
-
-	spin_lock_bh(&vport_config->mac_filter_list_lock);
-
-	f = idpf_find_mac_filter(vport_config, vport->default_mac_addr);
-	if (f) {
-		list_del(&f->list);
-		kfree(f);
-	}
-
-	spin_unlock_bh(&vport_config->mac_filter_list_lock);
-}
-
-/**
- * idpf_init_mac_addr - initialize mac address for vport
- * @vport: main vport structure
- * @netdev: pointer to netdev struct associated with this vport
- */
-static int idpf_init_mac_addr(struct idpf_vport *vport,
-			      struct net_device *netdev)
-{
-	struct idpf_netdev_priv *np = netdev_priv(netdev);
-	struct idpf_adapter *adapter = vport->adapter;
-	int err;
-
-	if (is_valid_ether_addr(vport->default_mac_addr)) {
-		eth_hw_addr_set(netdev, vport->default_mac_addr);
-		ether_addr_copy(netdev->perm_addr, vport->default_mac_addr);
-
-		return idpf_add_mac_filter(vport, np, vport->default_mac_addr,
-					   false);
-	}
-
-	if (!idpf_is_cap_ena(adapter, IDPF_OTHER_CAPS,
-			     VIRTCHNL2_CAP_MACFILTER)) {
-		dev_err(&adapter->pdev->dev,
-			"MAC address is not provided and capability is not set\n");
-
-		return -EINVAL;
-	}
-
-	eth_hw_addr_random(netdev);
-	err = idpf_add_mac_filter(vport, np, netdev->dev_addr, false);
-	if (err)
-		return err;
-
-	dev_info(&adapter->pdev->dev, "Invalid MAC address %pM, using random %pM\n",
-		 vport->default_mac_addr, netdev->dev_addr);
-	ether_addr_copy(vport->default_mac_addr, netdev->dev_addr);
-
-	return 0;
 }
 
 /**
