+++ drivers/net/ethernet/intel/idpf/idpf_virtchnl.c
@@ -4,102 +4,6 @@
 #include "idpf.h"
 #include "idpf_virtchnl.h"
 
-#define IDPF_VC_XN_MIN_TIMEOUT_MSEC	2000
-#define IDPF_VC_XN_DEFAULT_TIMEOUT_MSEC	(60 * 1000)
-#define IDPF_VC_XN_IDX_M		GENMASK(7, 0)
-#define IDPF_VC_XN_SALT_M		GENMASK(15, 8)
-#define IDPF_VC_XN_RING_LEN		U8_MAX
-
-/**
- * enum idpf_vc_xn_state - Virtchnl transaction status
- * @IDPF_VC_XN_IDLE: not expecting a reply, ready to be used
- * @IDPF_VC_XN_WAITING: expecting a reply, not yet received
- * @IDPF_VC_XN_COMPLETED_SUCCESS: a reply was expected and received,
- *				  buffer updated
- * @IDPF_VC_XN_COMPLETED_FAILED: a reply was expected and received, but there
- *				 was an error, buffer not updated
- * @IDPF_VC_XN_SHUTDOWN: transaction object cannot be used, VC torn down
- * @IDPF_VC_XN_ASYNC: transaction sent asynchronously and doesn't have the
- *		      return context; a callback may be provided to handle
- *		      return
- */
-enum idpf_vc_xn_state {
-	IDPF_VC_XN_IDLE = 1,
-	IDPF_VC_XN_WAITING,
-	IDPF_VC_XN_COMPLETED_SUCCESS,
-	IDPF_VC_XN_COMPLETED_FAILED,
-	IDPF_VC_XN_SHUTDOWN,
-	IDPF_VC_XN_ASYNC,
-};
-
-struct idpf_vc_xn;
-/* Callback for asynchronous messages */
-typedef int (*async_vc_cb) (struct idpf_adapter *, struct idpf_vc_xn *,
-			    const struct idpf_ctlq_msg *);
-
-/**
- * struct idpf_vc_xn - Data structure representing virtchnl transactions
- * @completed: virtchnl event loop uses that to signal when a reply is
- *	       available, uses kernel completion API
- * @state: virtchnl event loop stores the data below, protected by the
- *	   completion's lock.
- * @reply_sz: Original size of reply, may be > reply_buf.iov_len; it will be
- *	      truncated on its way to the receiver thread according to
- *	      reply_buf.iov_len.
- * @reply: Reference to the buffer(s) where the reply data should be written
- *	   to. May be 0-length (then NULL address permitted) if the reply data
- *	   should be ignored.
- * @async_handler: if sent asynchronously, a callback can be provided to handle
- *		   the reply when it's received
- * @vc_op: corresponding opcode sent with this transaction
- * @idx: index used as retrieval on reply receive, used for cookie
- * @salt: changed every message to make unique, used for cookie
- */
-struct idpf_vc_xn {
-	struct completion completed;
-	enum idpf_vc_xn_state state;
-	size_t reply_sz;
-	struct kvec reply;
-	async_vc_cb async_handler;
-	u32 vc_op;
-	u8 idx;
-	u8 salt;
-};
-
-/**
- * struct idpf_vc_xn_params - Parameters for executing transaction
- * @send_buf: kvec for send buffer
- * @recv_buf: kvec for recv buffer, may be NULL, must then have zero length
- * @timeout_ms: timeout to wait for reply
- * @async: send message asynchronously, will not wait on completion
- * @async_handler: If sent asynchronously, optional callback handler. The user
- *		   must be careful when using async handlers as the memory for
- *		   the recv_buf _cannot_ be on stack if this is async.
- * @vc_op: virtchnl op to send
- */
-struct idpf_vc_xn_params {
-	struct kvec send_buf;
-	struct kvec recv_buf;
-	int timeout_ms;
-	bool async;
-	async_vc_cb async_handler;
-	u32 vc_op;
-};
-
-/**
- * struct idpf_vc_xn_manager - Manager for tracking transactions
- * @ring: backing and lookup for transactions
- * @free_xn_bm: bitmap for free transactions
- * @xn_bm_lock: make bitmap access synchronous where necessary
- * @salt: used to make cookie unique every message
- */
-struct idpf_vc_xn_manager {
-	struct idpf_vc_xn ring[IDPF_VC_XN_RING_LEN];
-	DECLARE_BITMAP(free_xn_bm, IDPF_VC_XN_RING_LEN);
-	spinlock_t xn_bm_lock;
-	u8 salt;
-};
-
 /**
  * idpf_vid_to_vport - Translate vport id to vport pointer
  * @adapter: private data struct
@@ -107,7 +11,6 @@ struct idpf_vc_xn_manager {
  *
  * Returns vport matching v_id, NULL if not found.
  */
-static
 struct idpf_vport *idpf_vid_to_vport(struct idpf_adapter *adapter, u32 v_id)
 {
 	u16 num_max_vports = idpf_get_max_vports(adapter);
@@ -447,8 +350,8 @@ static void idpf_vc_xn_push_free(struct idpf_vc_xn_manager *vcxn_mngr,
  * >= @recv_buf.iov_len, but we never overflow @@recv_buf_iov_base). < 0 for
  * error.
  */
-static ssize_t idpf_vc_xn_exec(struct idpf_adapter *adapter,
-			       const struct idpf_vc_xn_params *params)
+ssize_t idpf_vc_xn_exec(struct idpf_adapter *adapter,
+			const struct idpf_vc_xn_params *params)
 {
 	const struct kvec *send_buf = &params->send_buf;
 	struct idpf_vc_xn *xn;
@@ -3465,226 +3368,4 @@ u32 idpf_get_vport_id(struct idpf_vport *vport)
 	vport_msg = vport->adapter->vport_params_recvd[vport->idx];
 
 	return le32_to_cpu(vport_msg->vport_id);
-}
-
-/**
- * idpf_mac_filter_async_handler - Async callback for mac filters
- * @adapter: private data struct
- * @xn: transaction for message
- * @ctlq_msg: received message
- *
- * In some scenarios driver can't sleep and wait for a reply (e.g.: stack is
- * holding rtnl_lock) when adding a new mac filter. It puts us in a difficult
- * situation to deal with errors returned on the reply. The best we can
- * ultimately do is remove it from our list of mac filters and report the
- * error.
- */
-static int idpf_mac_filter_async_handler(struct idpf_adapter *adapter,
-					 struct idpf_vc_xn *xn,
-					 const struct idpf_ctlq_msg *ctlq_msg)
-{
-	struct virtchnl2_mac_addr_list *ma_list;
-	struct idpf_vport_config *vport_config;
-	struct virtchnl2_mac_addr *mac_addr;
-	struct idpf_mac_filter *f, *tmp;
-	struct list_head *ma_list_head;
-	struct idpf_vport *vport;
-	u16 num_entries;
-	int i;
-
-	/* if success we're done, we're only here if something bad happened */
-	if (!ctlq_msg->cookie.mbx.chnl_retval)
-		return 0;
-
-	/* make sure at least struct is there */
-	if (xn->reply_sz < sizeof(*ma_list))
-		goto invalid_payload;
-
-	ma_list = ctlq_msg->ctx.indirect.payload->va;
-	mac_addr = ma_list->mac_addr_list;
-	num_entries = le16_to_cpu(ma_list->num_mac_addr);
-	/* we should have received a buffer at least this big */
-	if (xn->reply_sz < struct_size(ma_list, mac_addr_list, num_entries))
-		goto invalid_payload;
-
-	vport = idpf_vid_to_vport(adapter, le32_to_cpu(ma_list->vport_id));
-	if (!vport)
-		goto invalid_payload;
-
-	vport_config = adapter->vport_config[le32_to_cpu(ma_list->vport_id)];
-	ma_list_head = &vport_config->user_config.mac_filter_list;
-
-	/* We can't do much to reconcile bad filters at this point, however we
-	 * should at least remove them from our list one way or the other so we
-	 * have some idea what good filters we have.
-	 */
-	spin_lock_bh(&vport_config->mac_filter_list_lock);
-	list_for_each_entry_safe(f, tmp, ma_list_head, list)
-		for (i = 0; i < num_entries; i++)
-			if (ether_addr_equal(mac_addr[i].addr, f->macaddr))
-				list_del(&f->list);
-	spin_unlock_bh(&vport_config->mac_filter_list_lock);
-	dev_err_ratelimited(&adapter->pdev->dev, "Received error sending MAC filter request (op %d)\n",
-			    xn->vc_op);
-
-	return 0;
-
-invalid_payload:
-	dev_err_ratelimited(&adapter->pdev->dev, "Received invalid MAC filter payload (op %d) (len %zd)\n",
-			    xn->vc_op, xn->reply_sz);
-
-	return -EINVAL;
-}
-
-/**
- * idpf_add_del_mac_filters - Add/del mac filters
- * @vport: Virtual port data structure
- * @np: Netdev private structure
- * @add: Add or delete flag
- * @async: Don't wait for return message
- *
- * Returns 0 on success, error on failure.
- **/
-int idpf_add_del_mac_filters(struct idpf_vport *vport,
-			     struct idpf_netdev_priv *np,
-			     bool add, bool async)
-{
-	struct virtchnl2_mac_addr_list *ma_list __free(kfree) = NULL;
-	struct virtchnl2_mac_addr *mac_addr __free(kfree) = NULL;
-	struct idpf_adapter *adapter = np->adapter;
-	struct idpf_vc_xn_params xn_params = {};
-	struct idpf_vport_config *vport_config;
-	u32 num_msgs, total_filters = 0;
-	struct idpf_mac_filter *f;
-	ssize_t reply_sz;
-	int i = 0, k;
-
-	xn_params.vc_op = add ? VIRTCHNL2_OP_ADD_MAC_ADDR :
-				VIRTCHNL2_OP_DEL_MAC_ADDR;
-	xn_params.timeout_ms = IDPF_VC_XN_DEFAULT_TIMEOUT_MSEC;
-	xn_params.async = async;
-	xn_params.async_handler = idpf_mac_filter_async_handler;
-
-	vport_config = adapter->vport_config[np->vport_idx];
-	spin_lock_bh(&vport_config->mac_filter_list_lock);
-
-	/* Find the number of newly added filters */
-	list_for_each_entry(f, &vport_config->user_config.mac_filter_list,
-			    list) {
-		if (add && f->add)
-			total_filters++;
-		else if (!add && f->remove)
-			total_filters++;
-	}
-
-	if (!total_filters) {
-		spin_unlock_bh(&vport_config->mac_filter_list_lock);
-
-		return 0;
-	}
-
-	/* Fill all the new filters into virtchannel message */
-	mac_addr = kcalloc(total_filters, sizeof(struct virtchnl2_mac_addr),
-			   GFP_ATOMIC);
-	if (!mac_addr) {
-		spin_unlock_bh(&vport_config->mac_filter_list_lock);
-
-		return -ENOMEM;
-	}
-
-	list_for_each_entry(f, &vport_config->user_config.mac_filter_list,
-			    list) {
-		if (add && f->add) {
-			ether_addr_copy(mac_addr[i].addr, f->macaddr);
-			i++;
-			f->add = false;
-			if (i == total_filters)
-				break;
-		}
-		if (!add && f->remove) {
-			ether_addr_copy(mac_addr[i].addr, f->macaddr);
-			i++;
-			f->remove = false;
-			if (i == total_filters)
-				break;
-		}
-	}
-
-	spin_unlock_bh(&vport_config->mac_filter_list_lock);
-
-	/* Chunk up the filters into multiple messages to avoid
-	 * sending a control queue message buffer that is too large
-	 */
-	num_msgs = DIV_ROUND_UP(total_filters, IDPF_NUM_FILTERS_PER_MSG);
-
-	for (i = 0, k = 0; i < num_msgs; i++) {
-		u32 entries_size, buf_size, num_entries;
-
-		num_entries = min_t(u32, total_filters,
-				    IDPF_NUM_FILTERS_PER_MSG);
-		entries_size = sizeof(struct virtchnl2_mac_addr) * num_entries;
-		buf_size = struct_size(ma_list, mac_addr_list, num_entries);
-
-		if (!ma_list || num_entries != IDPF_NUM_FILTERS_PER_MSG) {
-			kfree(ma_list);
-			ma_list = kzalloc(buf_size, GFP_ATOMIC);
-			if (!ma_list)
-				return -ENOMEM;
-		} else {
-			memset(ma_list, 0, buf_size);
-		}
-
-		ma_list->vport_id = cpu_to_le32(np->vport_id);
-		ma_list->num_mac_addr = cpu_to_le16(num_entries);
-		memcpy(ma_list->mac_addr_list, &mac_addr[k], entries_size);
-
-		xn_params.send_buf.iov_base = ma_list;
-		xn_params.send_buf.iov_len = buf_size;
-		reply_sz = idpf_vc_xn_exec(adapter, &xn_params);
-		if (reply_sz < 0)
-			return reply_sz;
-
-		k += num_entries;
-		total_filters -= num_entries;
-	}
-
-	return 0;
-}
-
-/**
- * idpf_set_promiscuous - set promiscuous and send message to mailbox
- * @adapter: Driver specific private structure
- * @config_data: Vport specific config data
- * @vport_id: Vport identifier
- *
- * Request to enable promiscuous mode for the vport. Message is sent
- * asynchronously and won't wait for response.  Returns 0 on success, negative
- * on failure;
- */
-int idpf_set_promiscuous(struct idpf_adapter *adapter,
-			 struct idpf_vport_user_config_data *config_data,
-			 u32 vport_id)
-{
-	struct idpf_vc_xn_params xn_params = {};
-	struct virtchnl2_promisc_info vpi;
-	ssize_t reply_sz;
-	u16 flags = 0;
-
-	if (test_bit(__IDPF_PROMISC_UC, config_data->user_flags))
-		flags |= VIRTCHNL2_UNICAST_PROMISC;
-	if (test_bit(__IDPF_PROMISC_MC, config_data->user_flags))
-		flags |= VIRTCHNL2_MULTICAST_PROMISC;
-
-	vpi.vport_id = cpu_to_le32(vport_id);
-	vpi.flags = cpu_to_le16(flags);
-
-	xn_params.vc_op = VIRTCHNL2_OP_CONFIG_PROMISCUOUS_MODE;
-	xn_params.timeout_ms = IDPF_VC_XN_DEFAULT_TIMEOUT_MSEC;
-	xn_params.send_buf.iov_base = &vpi;
-	xn_params.send_buf.iov_len = sizeof(vpi);
-	/* setting promiscuous is only ever done asynchronously */
-	xn_params.async = true;
-	reply_sz = idpf_vc_xn_exec(adapter, &xn_params);
-
-	return reply_sz < 0 ? reply_sz : 0;
 }
